AI-Gateway/labs/master-lab/requirements.txt
# ============================================================================
# Master AI Gateway Lab - Python Dependencies
# ============================================================================
# Complete set of dependencies for running all exercises in master-ai-gateway.ipynb
# ============================================================================

# ----------------------------------------------------------------------------
# Environment & Configuration
# ----------------------------------------------------------------------------
python-dotenv>=1.0.0

# ----------------------------------------------------------------------------
# Core Azure SDKs
# ----------------------------------------------------------------------------
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0
azure-mgmt-resource>=23.0.0
azure-mgmt-apimanagement>=4.0.0
azure-mgmt-cognitiveservices>=13.5.0
azure-mgmt-cosmosdb>=9.0.0
azure-core>=1.29.0
azure-cosmos>=4.5.0
azure-monitor-query>=1.2.0

# ----------------------------------------------------------------------------
# Azure AI & OpenAI
# ----------------------------------------------------------------------------
openai>=1.12.0
azure-ai-inference>=1.0.0b1
azure-search-documents>=11.4.0

# ----------------------------------------------------------------------------
# Model Context Protocol (MCP)
# ----------------------------------------------------------------------------
mcp>=0.9.0
httpx>=0.27.0

# ----------------------------------------------------------------------------
# Agent Frameworks
# ----------------------------------------------------------------------------
# Microsoft Agent Framework (Azure AI Agent Service)
#agent-framework>=1.0.0b1

# Semantic Kernel
semantic-kernel>=1.0.0

# AutoGen (Legacy 0.2.x - Conversational Agents)
# Requires Python >=3.8,<3.12 (use .venv-py311)
pyautogen~=0.2.0

# ----------------------------------------------------------------------------
# Data Processing & Analysis
# ----------------------------------------------------------------------------
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0
matplotlib>=3.8.0
Pillow>=10.0.0

# ----------------------------------------------------------------------------
# Web & API
# ----------------------------------------------------------------------------
requests>=2.31.0
httpx>=0.27.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiohttp>=3.9.0
sse-starlette>=1.8.0

# ----------------------------------------------------------------------------
# Authentication & Security
# ----------------------------------------------------------------------------
PyJWT>=2.8.0
cryptography>=42.0.0
msal>=1.26.0
python-multipart>=0.0.6

# ----------------------------------------------------------------------------
# Environment & Configuration
# ----------------------------------------------------------------------------
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ----------------------------------------------------------------------------
# Windows Support (NOT compatible with WSL - installed conditionally in Cell 007)
# ----------------------------------------------------------------------------
# pywin32>=306  # Windows-only, not available on Linux/WSL

# ----------------------------------------------------------------------------
# Redis (for caching examples)
# ----------------------------------------------------------------------------
redis>=5.0.0

# ----------------------------------------------------------------------------
# Async & Concurrency
# ----------------------------------------------------------------------------
nest-asyncio>=1.6.0

# ----------------------------------------------------------------------------
# Jupyter Notebook Support
# ----------------------------------------------------------------------------
ipython>=8.20.0
ipykernel>=6.29.0
jupyter>=1.0.0
nbconvert>=7.14.0
nbformat>=5.9.0
papermill>=2.4.0

# ----------------------------------------------------------------------------
# Development & Testing
# ----------------------------------------------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
isort>=5.13.0
pylint>=3.0.0
mypy>=1.8.0

# ----------------------------------------------------------------------------
# Additional Utilities
# ----------------------------------------------------------------------------
tqdm>=4.66.0
tabulate>=0.9.0
colorama>=0.4.6
